Added phpunit.xml.dist template file
This is temporary, and it loads a autoload.php file from a Symfony
Standard Edition, the one used to initially test this bundle. In
order for the tests to work, you should create a phpunit.xml file
based on this and add the correct autoload.php (which is probably
going to be the one from composer).